+++ src/tag/parser/compile/scenario/post_compile.cpp
@@ -31,7 +31,7 @@ namespace Invader::Parser {
         auto bsp_count = bsp_data.size();
 
         // Determine which BSP light fixtures and scenery are in
-        #define FIND_BSP_INDICES_FOR_OBJECT_ARRAY(array_type, objects) { \
+        #define FIND_BSP_INDICES_FOR_OBJECT_ARRAY(array_type, objects, palette_type, palette, type_name) { \
             std::size_t object_count = this->objects.size(); \
             if(object_count) { \
                 auto &object_struct = workload.structs[*scenario_struct.resolve_pointer(&scenario_data.objects.pointer)]; \
@@ -40,12 +40,18 @@ namespace Invader::Parser {
                     std::uint32_t bsp_indices = 0; \
                     auto &object = object_array[o]; \
                     if(object.type != NULL_INDEX) { \
+                        auto &type = this->palette[object.type].name; \
+                        auto &object_bounding_offset = reinterpret_cast<Object::struct_little *>(workload.structs[*workload.tags[type.tag_id.index].base_struct].data.data())->bounding_offset; \
+                        auto position_to_check = object.position + object_bounding_offset; \
                         for(std::size_t b = 0; b < bsp_count; b++) { \
                             /* Check if we're inside this BSP */ \
                             auto &bsp = bsp_data[b]; \
-                            if(bsp.check_if_point_inside_bsp(object.position)) { \
+                            if(bsp.check_if_point_inside_bsp(position_to_check)) { \
                                 bsp_indices |= 1 << b; \
                             } \
+                        } \
+                        if(bsp_indices == 0) { \
+                            REPORT_ERROR_PRINTF(workload, ERROR_TYPE_WARNING, tag_index, type_name " spawn #%zu was found in 0 BSPs, so it will not spawn", o); \
                         } \
                     } \
                     /* Set to the result */ \
@@ -54,8 +60,8 @@ namespace Invader::Parser {
             } \
         }
 
-        FIND_BSP_INDICES_FOR_OBJECT_ARRAY(ScenarioScenery, scenery);
-        FIND_BSP_INDICES_FOR_OBJECT_ARRAY(ScenarioLightFixture, light_fixtures);
+        FIND_BSP_INDICES_FOR_OBJECT_ARRAY(ScenarioScenery, scenery, ScenarioSceneryPalette, scenery_palette, "Scenery");
+        FIND_BSP_INDICES_FOR_OBJECT_ARRAY(ScenarioLightFixture, light_fixtures, ScenarioLightFixturePalette, light_fixture_palette, "Light fixture");
         
         #undef FIND_BSP_INDICES_FOR_OBJECT_ARRAY